requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-filter>=24.3,<25.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
